frontend/requirements.txt
gunicorn
flask
gevent
Flask-Migrate
Flask-SQLAlchemy
Flask-Mail
Flask-Login
Flask-Moment
Flask-WTF
Flask-HTTPAuth
email_validator
SQLAlchemy
python-dotenv
PyJWT
pytest
pika
requests
pandas
openpyxl
